requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
neo4j==5.23.0
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1
langdetect==1.0.9
regex==2024.5.15
pymupdf==1.24.9
pdfplumber==0.11.4
unidecode==1.3.8
tqdm==4.66.5
networkx==3.3
langchain-community==0.3.29
langchain==0.3.27
pypdf==6.0.0
langgraph==0.6.6